Event handler lib rewrite (#5)
* Event handler lib rewrite

* Addressing own CR comments
